+++ src/lib/stategraph.rs
@@ -200,7 +200,7 @@ pub fn state_exists(grm: &YaccGrammar, is: &Itemset, nt: &str, prod_off: usize, 
 
     let ab_prod_off = grm.nonterm_to_prods(grm.nonterm_idx(nt).unwrap())[prod_off];
     let ctx = &is.items[&(ab_prod_off, dot.into())];
-    for i in 0..grm.terms_len() {
+    for i in 0..grm.terms_len() as usize {
         let bit = ctx[i];
         let mut found = false;
         for t in la.iter() {
